Refactor: 제품 API 인증 체크 예외 클래스의 위치를 옮긴다
