+++ src/main/java/com/flab/sooldama/global/auth/AuthInterceptor.java
@@ -1,6 +1,6 @@
 package com.flab.sooldama.global.auth;
 
-import com.flab.sooldama.global.exception.AuthenticationFailException;
+import com.flab.sooldama.domain.product.exception.AuthenticationFailException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
