+++ src/test/java/com/flab/sooldama/domain/product/api/ProductApiTest.java
@@ -12,7 +12,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.flab.sooldama.domain.product.dto.response.ProductResponse;
 import com.flab.sooldama.domain.product.exception.ProductNotFoundException;
 import com.flab.sooldama.domain.product.service.ProductService;
-import com.flab.sooldama.global.exception.AuthenticationFailException;
+import com.flab.sooldama.domain.product.exception.AuthenticationFailException;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
